ospf: advertise configured flex-algos in SR-Algorithm TLV (v2)
Phase 3a — first on-the-wire flex-algo behavior. The OSPFv2 Router
Information LSA's SR-Algorithm TLV (type 8) was hardcoded to Algo::Spf;
it now lists SPF plus every configured Flexible Algorithm (RFC 9350
§6), so peers see this router participate.

- Move sr_algorithms() to crate::flex_algo (it operates only on the
  shared FlexAlgoConfig + packet_utils::Algo, which both isis-packet
  and ospf-packet re-export — one helper serves both IGPs). isis
  re-exports it; no behavior change there.
- router_info_lsa_build takes the algo list; router_info_lsa_originate
  passes flex_algo::sr_algorithms(&self.flex_algo).

Still gated on segment-routing/mpls (the RI LSA only originates with SR
enabled); FAD origination + per-link ASLA follow. New test asserts the
TLV carries the configured algos. Full suite (770) + clippy + fmt clean.
