+++ zebra-rs/src/isis/flex_algo.rs
@@ -20,6 +20,7 @@ use super::srlg::SrlgGroup;
 // IS-IS callback shims below stay here.
 pub use crate::flex_algo::{
     FadMetricType, FlexAlgoConfig, FlexAlgoEntry, link_passes_fad, local_link_affinity,
+    sr_algorithms,
 };
 
 /// Extract per-algorithm Prefix-SIDs from a peer-advertised Ext IP-
@@ -123,25 +124,6 @@ pub fn build_per_algo_prefix_sids(map: &BTreeMap<u8, SidLabelValue>) -> Vec<Isis
             sid: sid.clone(),
         })
         .collect()
-}
-
-/// Algorithms this router advertises in the SR Algorithm sub-TLV
-/// (RFC 8667 §3.2, sub-TLV 19). `Algo::Spf` (algo 0) is always
-/// present; every flex-algo entry in `fa.config` is added as
-/// `Algo::FlexAlgo(N)` in sorted order.
-///
-/// Required by RFC 9350 §5.2: a router that originates a FAD or
-/// participates in a flex-algo MUST list that algo here. The
-/// configuration model treats *any* entry in `flex_algo.config` as
-/// participation — `advertise_definition` controls FAD origination,
-/// not participation.
-pub fn sr_algorithms(fa: &FlexAlgoConfig) -> Vec<Algo> {
-    let mut algos = Vec::with_capacity(1 + fa.config.len());
-    algos.push(Algo::Spf);
-    for &n in fa.config.keys() {
-        algos.push(Algo::FlexAlgo(n));
-    }
-    algos
 }
 
 /// Build the FAD sub-TLVs (RFC 9350 §5.1) this router will originate
@@ -567,32 +549,6 @@ mod tests {
             }
             _ => panic!("expected AdminGrp"),
         }
-    }
-
-    #[test]
-    fn sr_algorithms_lists_spf_plus_every_configured_algo() {
-        let mut fa = FlexAlgoConfig::new("/router/isis/flex-algo");
-        // No flex-algos yet — should yield exactly Algo::Spf.
-        assert_eq!(sr_algorithms(&fa), vec![Algo::Spf]);
-
-        // Add two flex-algos via any leaf write (priority is fine —
-        // participation is implied by the entry existing, not by
-        // advertise_definition).
-        for algo in ["129", "128"] {
-            fa.exec(
-                "/router/isis/flex-algo/priority".into(),
-                args(&[algo, "200"]),
-                ConfigOp::Set,
-            )
-            .unwrap();
-            fa.commit();
-        }
-        // BTreeMap iterates sorted, so flex-algos appear in 128, 129
-        // order after Spf.
-        assert_eq!(
-            sr_algorithms(&fa),
-            vec![Algo::Spf, Algo::FlexAlgo(128), Algo::FlexAlgo(129)]
-        );
     }
 
     #[test]
